apps/api/pyproject.toml
[project]
name = "adobe-pipeline-api"
version = "0.1.0"
description = "FastAPI backend for Adobe creative automation pipeline"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.4",
    "campaign-schema",
    "fastapi>=0.115.0",
    "uvicorn>=0.35.0",
    "pydantic>=2.11.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.28.0",
    "pyjwt[crypto]>=2.10.0",
    "supabase>=2.15.0",
    "anthropic>=0.50.0",
    "lumaai>=1.0.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "httpx>=0.28.1",
    "ruff>=0.8.0",
]

[tool.uv.sources]
campaign-schema = { path = "../../packages/campaign-schema", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
